+++ Course/InputRect4x5/TaskRect4x5.cs
@@ -1,4 +1,4 @@
-using Real = float;
+using Real = double;
 
 class TaskRect4x5 : TaskFuncs
 {
@@ -78,6 +78,6 @@ class TaskRect4x5 : TaskFuncs
             3 => (Real)(Math.Exp(x+y) + x),
             _ => throw new ArgumentException("Некорректный номер условия"),
         };
-    }    
+    }
 
 }
